fix annoying issue
so: intuitively, the order of the items in a uniform block need to be
correct. yet unintuitively, there is absolutely zero error messages if
they are out of order. I probably could check the logs or something, but
I feel like that should probably be a crash instead? although we are
just straight up turning our Uniform struct into bytes and throwing it
at the gpu, so I don't even know how you might check something like
that. anyway. fun.